Adds method to perform extracting to intials in the EventViewModel
